pyproject.toml
[tool.poetry]
name = "tweeter"
version = "0.1.0"
description = ""
authors = ["chifarol <ilodigwechinaza@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "^3.5.2"
certifi = "^2022.6.15"
cffi = "^1.15.1"
cloudinary = "^1.29.0"
cryptography = "^37.0.4"
Django = "^4.0.6"
django-environ = "^0.9.0"
django-rest-knox = "^4.2.0"
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"
pycparser = "^2.21"
python-dotenv = "^0.20.0"
pytz = "^2022.1"
six = "^1.16.0"
sqlparse = "^0.4.2"
tzdata = "^2022.1"
urllib3 = "^1.26.11"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
